requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==23.2.1
albumentations==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
basicsr==1.4.2
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.1
customtkinter==5.2.0
cycler==0.12.1
Cython==3.0.12
darkdetect==0.8.0
easydict==1.13
exceptiongroup==1.2.2
facexlib==0.3.0
fastapi==0.115.11
ffmpeg==1.4
ffmpy==0.5.0
filelock==3.18.0
filterpy==1.4.5
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
future==1.0.0
gast==0.4.0
gdown==5.2.0
gfpgan==1.3.8
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==5.22.0
gradio_client==1.8.0
groovy==0.1.2
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
imageio==2.37.0
insightface==0.7.3
Jinja2==3.1.6
joblib==1.4.2
keras==2.13.1
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
lmdb==1.6.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
numba==0.61.0
numpy==1.24.3
oauthlib==3.2.2
onnx==1.14.0
onnxruntime-gpu==1.15.1
opencv-python==4.8.0.74
opencv-python-headless==4.11.0.86
opennsfw2==0.10.2
opt_einsum==3.4.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
Pillow==10.0.0
prettytable==3.15.1
protobuf==4.23.4
psutil==5.9.5
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
qudida==0.0.4
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
ruff==0.11.1
safehttpx==0.1.6
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
tb-nightly==2.20.0a20250317
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.6.0
tifffile==2025.3.13
tk==0.1.0
tkinterdnd2==0.3.0
tomlkit==0.13.2
torch==2.1.2
torchvision==0.16.2+cu118
tqdm==4.65.0
triton==2.1.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
wrapt==1.17.2
yapf==0.43.0
